requirements.txt
python_socketio==5.6.0
aiohttp==3.8.1
pyyaml
selenium
python-dotenv